services/frontend/requirements.txt
flask
configobj
passlib
Flask-WTF
requests
pyjwt
